Changed data type of in_flight_request_count to double in in_flight_requests stream

+++ components/org.apache.stratos.load.balancer.common/src/main/java/org/apache/stratos/load/balancer/common/statistics/publisher/WSO2CEPInFlightRequestPublisher.java
@@ -47,7 +47,7 @@ public class WSO2CEPInFlightRequestPublisher extends WSO2CEPStatisticsPublisher 
             // Payload definition
             payloadData.add(new Attribute("cluster_id", AttributeType.STRING));
             payloadData.add(new Attribute("network_partition_id", AttributeType.STRING));
-            payloadData.add(new Attribute("in_flight_request_count", AttributeType.INT));
+            payloadData.add(new Attribute("in_flight_request_count", AttributeType.DOUBLE));
             streamDefinition.setPayloadData(payloadData);
             return streamDefinition;
         } catch (Exception e) {
@@ -71,7 +71,7 @@ public class WSO2CEPInFlightRequestPublisher extends WSO2CEPStatisticsPublisher 
         // Payload values
         payload.add(clusterId);
         payload.add(networkPartitionId);
-        payload.add(inFlightRequestCount);
+        payload.add((double)inFlightRequestCount);
         super.publish(payload.toArray());
     }
 }
